Removing prefixHandler
